intent_engine/requirements.txt
Django>=4.2
djangorestframework
PyYAML
